+++ Libplanet.Benchmarks/ProposeBlock.cs
@@ -105,31 +105,31 @@ namespace Libplanet.Benchmarks
         [Benchmark]
         public void ProposeBlockEmpty()
         {
-            _block = _blockChain.ProposeBlock(_privateKey, lastCommit: _lastCommit);
+            _block = _blockChain.ProposeBlock(_privateKey, _lastCommit);
         }
 
         [Benchmark]
         public void ProposeBlockOneTransactionNoAction()
         {
-            _block = _blockChain.ProposeBlock(_privateKey, lastCommit: _lastCommit);
+            _block = _blockChain.ProposeBlock(_privateKey, _lastCommit);
         }
 
         [Benchmark]
         public void ProposeBlockTenTransactionsNoAction()
         {
-            _block = _blockChain.ProposeBlock(_privateKey, lastCommit: _lastCommit);
+            _block = _blockChain.ProposeBlock(_privateKey, _lastCommit);
         }
 
         [Benchmark]
         public void ProposeBlockOneTransactionWithActions()
         {
-            _block = _blockChain.ProposeBlock(_privateKey, lastCommit: _lastCommit);
+            _block = _blockChain.ProposeBlock(_privateKey, _lastCommit);
         }
 
         [Benchmark]
         public void ProposeBlockTenTransactionsWithActions()
         {
-            _block = _blockChain.ProposeBlock(_privateKey, lastCommit: _lastCommit);
+            _block = _blockChain.ProposeBlock(_privateKey, _lastCommit);
         }
     }
 }