+++ Libplanet.Net.Tests/Consensus/Context/MessageLogTest.cs
@@ -26,9 +26,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
         {
             _blockChain = TestUtils.CreateDummyBlockChain(
                 new MemoryStoreFixture(TestUtils.Policy.BlockAction));
-            var block = _blockChain.ProposeBlock(
-                TestUtils.PrivateKeys[1],
-                DateTimeOffset.UtcNow);
+            var block = _blockChain.ProposeBlock(TestUtils.PrivateKeys[1]);
             _lastCommit = TestUtils.CreateBlockCommit(block);
             _messageLog = new MessageLog(2, TestUtils.ValidatorSet);
             _blockChain.Append(block, TestUtils.CreateBlockCommit(block));
@@ -51,10 +49,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
         [Fact]
         public void CannotAddInvalidProposer()
         {
-            var block = _blockChain.ProposeBlock(
-                TestUtils.PrivateKeys[0],
-                DateTimeOffset.UtcNow,
-                lastCommit: _lastCommit);
+            var block = _blockChain.ProposeBlock(TestUtils.PrivateKeys[0], _lastCommit);
             var proposal0 = new ConsensusProposalMsg(new ProposalMetadata(
                 2,
                 0,
@@ -87,10 +82,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
         [Fact]
         public void CannotAddMultipleProposalsPerRound()
         {
-            var block = _blockChain.ProposeBlock(
-                TestUtils.PrivateKeys[0],
-                DateTimeOffset.UtcNow,
-                lastCommit: _lastCommit);
+            var block = _blockChain.ProposeBlock(TestUtils.PrivateKeys[0], _lastCommit);
             var proposal0 = new ConsensusProposalMsg(new ProposalMetadata(
                 2,
                 0,
@@ -152,10 +144,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
         [Fact]
         public void GetValidatorsCount()
         {
-            var block = _blockChain.ProposeBlock(
-                TestUtils.PrivateKeys[0],
-                DateTimeOffset.UtcNow,
-                lastCommit: _lastCommit);
+            var block = _blockChain.ProposeBlock(TestUtils.PrivateKeys[0], _lastCommit);
             var proposal = new ConsensusProposalMsg(new ProposalMetadata(
                 2,
                 0,
@@ -182,10 +171,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
         [Fact]
         public void GetBlockCommit()
         {
-            var block = _blockChain.ProposeBlock(
-                TestUtils.PrivateKeys[0],
-                DateTimeOffset.UtcNow,
-                lastCommit: _lastCommit);
+            var block = _blockChain.ProposeBlock(TestUtils.PrivateKeys[0], _lastCommit);
             var proposal = new ConsensusProposalMsg(new ProposalMetadata(
                 2,
                 0,